requirements.txt
appdirs==1.4.3
argon2-cffi==20.1.0
asgiref==3.3.4
async-generator==1.10
attrs==21.2.0
backcall==0.2.0
bleach==3.3.0
CacheControl==0.12.6
cachetools==4.2.2
certifi==2019.11.28
cffi==1.14.5
chardet==3.0.4
click==8.0.1
colorama==0.4.3
contextlib2==0.6.0
decorator==5.0.9
defusedxml==0.7.1
distlib==0.3.0
distro==1.4.0
entrypoints==0.3
fastapi==0.65.2
h11==0.12.0
h3==3.7.3
html5lib==1.0.1
httptools==0.2.0
idna==2.8
ipaddr==2.2.0
ipykernel==5.5.5
ipython==7.24.1
ipython-genutils==0.2.0
jedi==0.18.0
Jinja2==3.0.1
jsonschema==3.2.0
jupyter-client==6.1.12
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
lockfile==0.12.2
MarkupSafe==2.0.1
matplotlib-inline==0.1.2
mistune==0.8.4
msgpack==0.6.2
nbclient==0.5.3
nbconvert==6.0.7
nbformat==5.1.3
nest-asyncio==1.5.1
notebook==6.4.0
numpy==1.20.3
packaging==20.3
pandocfilters==1.4.3
parso==0.8.2
pep517==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
progress==1.5
prometheus-client==0.11.0
prompt-toolkit==3.0.19
ptyprocess==0.7.0
pycparser==2.20
pydantic==1.8.2
Pygments==2.9.0
pyparsing==2.4.6
pyrsistent==0.17.3
python-dateutil==2.8.1
python-dotenv==0.18.0
pytoml==0.1.21
PyYAML==5.4.1
pyzmq==22.1.0
redis==3.5.3
requests==2.22.0
retrying==1.3.3
scipy==1.6.3
Send2Trash==1.5.0
six==1.14.0
starlette==0.14.2
terminado==0.10.1
testpath==0.5.0
tornado==6.1
traitlets==5.0.5
typing-extensions==3.10.0.0
urllib3==1.25.8
uvicorn==0.14.0
uvloop==0.15.2
watchgod==0.7
wcwidth==0.2.5
webencodings==0.5.1
websockets==9.1
